requirements.txt
python-dotenv
requests
tabulate
string